Minor fixes and changes
