+++ src/components/SearchProducts/SearchProducts.styles.ts
