+++ src/components/SearchProducts/index.ts
