+++ src/pages/SearchProductsPage.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { Container } from '@material-ui/core';
-import NavBar from '../components/NavBar';
-import SearchProducts from '../components/HeroSearchProducts';
+
+import SearchProducts from '../components/SearchProducts';
 
 const SearchProductsPage: React.FC = () => {
   return (
     <>
       <Container maxWidth="lg" disableGutters>
-        <NavBar />
         <SearchProducts />
       </Container>
     </>
